feat:#18 테스트 결과 삭제 알람 추가

+++ src/pages/Result.jsx
@@ -1,6 +1,7 @@
 import { useState, useEffect } from "react";
 import { deleteTestResult, getTestResults } from "@api/testResults";
 import ResultCard from "@components/result/ResultCard";
+import Swal from "sweetalert2";
 
 const Result = () => {
   const [isLoading, setIsLoading] = useState(true);
@@ -22,11 +23,35 @@ const Result = () => {
 
   const handleDelete = async (id) => {
     try {
-      const { nickname } = await deleteTestResult(id);
-      setTestResults((prev) => prev.filter((result) => result.id !== id));
-      alert(`${nickname}의 글이 삭제됐습니다.`);
+      const result = await Swal.fire({
+        icon: "warning",
+        title: "결과 삭제 확인",
+        text: "테스트 결과를 삭제하시겠습니까?",
+        showCancelButton: true,
+        confirmButtonText: "예",
+        cancelButtonText: "아니오",
+        confirmButtonColor: "#a5b4fc",
+        cancelButtonColor: "#fc7531",
+      });
+      if (result.isConfirmed)
+        if (result.isConfirmed) {
+          const { nickname } = await deleteTestResult(id);
+          setTestResults((prev) => prev.filter((result) => result.id !== id));
+          await Swal.fire({
+            title: "삭제 완료",
+            text: `${nickname}님의 테스트 결과를 삭제했습니다.`,
+            icon: "success",
+            confirmButtonColor: "#a5b4fc",
+          });
+        }
     } catch (e) {
       console.error(e);
+      Swal.fire({
+        title: "삭제 오류",
+        text: e.response.data.message,
+        icon: "error",
+        confirmButtonColor: "#a5b4fc",
+      });
     }
   };
